chore(tgi): refactor generator to separate base class
