+++ text-generation-inference/server/text_generation_server/generator.py
@@ -2,9 +2,8 @@ import copy
 import logging
 import os
 import time
-from abc import ABC
 from enum import Enum
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Tuple
 
 import torch
 import torch_xla.core.xla_model as xm
@@ -15,6 +14,7 @@ from transformers.generation import GenerationConfig
 from optimum.tpu import AutoModelForCausalLM
 from optimum.tpu.generation import TokenSelector
 
+from .generator_base import Generator
 from .pb.generate_pb2 import (
     Batch,
     CachedBatch,
@@ -30,63 +30,6 @@ from .pb.generate_pb2 import (
 # Disable optimum-tpu warnings as it seems to block the server after a while
 optimum_logger = logging.getLogger("optimum.tpu")
 optimum_logger.setLevel("CRITICAL")
-
-
-class Generator(ABC):
-    """An abstract class to represent the workhorse behind TextGenerationService.
-
-    Ideally, it should not rely on protobuf constructs, but in a first step it does.
-    Implementations would typically need a model and a tokenizer to implement the Generator methods.
-    """
-
-    @property
-    def info(self) -> InfoResponse:
-        """This should simply return the expected InfoResponse"""
-        raise NotImplementedError
-
-    def warmup(self, batch: Batch) -> int:
-        """Verify if the hardware can support the target load.
-
-        Args:
-            batch (`Batch`):
-                A batch corresponding to the maximum number of concurrent requests.
-
-        Return:
-            The maximum number of tokens the model supports.
-        """
-        raise NotImplementedError
-
-    def prefill(self, batch: Batch) -> Tuple[List[Generation], CachedBatch]:
-        """Prefill is called whenever new requests need to be added.
-
-        When this method returns successfully, a decode method will follow
-        with both the current and newly prefilled batch(es).
-
-        Args:
-            batch (`Batch`):
-                A batch containing the new requests.
-
-        Return:
-            A list of `Generation` for each request and a `CachedBatch` containing all pending requests.
-        """
-        raise NotImplementedError
-
-    def decode(self, batches: List[Batch]) -> Tuple[List[Generation], CachedBatch]:
-        """Decode after a prefill or another decode."""
-        raise NotImplementedError
-
-    def filter(self, batch_id: int, request_ids: List[int]) -> CachedBatch:
-        """Remove requests that are not listed from the specified batch"""
-        raise NotImplementedError
-
-    def clear(self):
-        """Remove all requests from the generator"""
-        raise NotImplementedError
-
-    @classmethod
-    def from_pretrained(cls, model_id: str, revision: Optional[str]):
-        """Factory method "a la transformers" """
-        raise NotImplementedError
 
 
 class Slot:
